requirements.txt
flask==3.0.2
flask-cors==4.0.0
google-generativeai==0.3.2
markdown==3.5.2
python-dotenv==1.0.1
streamlit==1.31.1
pandas==2.2.0
emoji==2.10.1 